remove old swap tool code
